+++ driver-core/CHANGELOG.rst
@@ -18,6 +18,7 @@ CHANGELOG
 - [improvement] Log datacenter name in DCAware policy's init when it is explicitly provided
   (JAVA-423)
 - [improvement] Shuffle the replicas in TokenAwarePolicy.newQueryPlan (JAVA-504)
+- [improvement] Make schema agreement wait tuneable (JAVA-507)
 
 
 2.0.6: